ideogram/ideogram-mcp/pyproject.toml
[project]
name = "ideogram-mcp"
version = "0.1.0"
description = "MCP server for Ideogram 3.0 API"
authors = [
    { name = "Petr", email = "petr@example.com" }
]
dependencies = [
    "mcp[cli]>=1.1.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0"
]
requires-python = ">=3.10"

[project.scripts]
ideogram-mcp = "ideogram_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
